+++ shared/types.ts
@@ -8,16 +8,6 @@ export type DomainScores = {
   verb: number;
 };
 
-export const blankScores: DomainScores = {
-  evoc: 0,
-  brev: 0,
-  grep: 0,
-  goog: 0,
-  pron: 0,
-  spel: 0,
-  verb: 0,
-};
-
 export type DomainAssessment = {
   domain: string;
   isPossible: boolean;
@@ -26,36 +16,4 @@ export type DomainAssessment = {
   scores?: DomainScores;
   status?: string;
   error?: string;
-};
-
-export const exampleValid: DomainAssessment = {
-  domain: "example.com",
-  isPossible: true,
-  isAvailable: true,
-  isCheap: true,
-  scores: blankScores,
-};
-
-export const exampleExpensive: DomainAssessment = {
-  domain: "table.com",
-  isPossible: true,
-  isAvailable: true,
-  isCheap: false,
-  scores: blankScores,
-};
-
-export const exampleUnavailable: DomainAssessment = {
-  domain: "google.com",
-  isPossible: true,
-  isAvailable: false,
-  isCheap: false,
-  scores: blankScores,
-};
-
-export const exampleImpossible: DomainAssessment = {
-  domain: "kandy.kettle",
-  isPossible: false,
-  isAvailable: false,
-  isCheap: false,
-  scores: blankScores,
 };
